Move individual classes to separate header/cpp files
