+++ mdr_perception/mdr_cloud_object_detection/ros/include/mdr_cloud_object_detection/ObjectFilterParams.h
@@ -1,0 +1,16 @@
+#ifndef MDR_CLOUD_OBJECT_DETECTION_OBJECT_FILTERING_PARAMS
+#define MDR_CLOUD_OBJECT_DETECTION_OBJECT_FILTERING_PARAMS
+
+namespace mdr_cloud_object_detection
+{
+    /*!
+    * @brief struct containing parameters necessary for filtering the detected objects
+    */
+    struct ObjectFilterParams
+    {
+        float mMaxBboxEdgeLength = 0.15f;
+        float mMinDistFromOccupiedCell = 0.1f;
+    };
+}
+
+#endif // MDR_CLOUD_OBJECT_DETECTION_OBJECT_FILTERING_PARAMS
