+++ mdr_perception/mdr_cloud_object_detection/ros/include/mdr_cloud_object_detection/Utils.h
@@ -1,0 +1,5 @@
+#include "mdr_cloud_object_detection/EuclideanClusterParams.h"
+#include "mdr_cloud_object_detection/ObjectCacheParams.h"
+#include "mdr_cloud_object_detection/ObjectFilterParams.h"
+
+#include "mdr_cloud_object_detection/PassThroughVoxelFilter.h"
