--- mdr_perception/mdr_cloud_object_detection/ros/include/mdr_cloud_object_detection/cloud_object_detection.h
@@ -1,1 +1,0 @@
-#include<iostream>
